+++ src/modules/transactions/transactions.routes.ts
@@ -1,12 +1,17 @@
 import { Router } from "express";
 import { validate } from "../../middlewares/validate.middleware.js";
 import { authMiddleware } from "../../middlewares/auth.middleware.js";
-import { createTransactionSchema } from "./transactions.schema.js";
+import {
+  createTransactionSchema,
+  updateTransactionSchema,
+} from "./transactions.schema.js";
 import {
   createTransactionController,
   getTransactionByIdController,
   getTransactionsController,
+  updateTransactionController,
 } from "./transactions.controller.js";
+import { idParamsSchema } from "../../shared/schemas/id.schema.js";
 
 const router = Router();
 
@@ -26,5 +31,17 @@ router.get("/", authMiddleware, getTransactionsController);
 // GET TRANSACTION BY ID
 
 router.get("/:id", authMiddleware, getTransactionByIdController);
+
+// UPDATE TRANSACTION
+
+router.patch(
+  "/:id",
+  authMiddleware,
+  validate({
+    params: idParamsSchema,
+    body: updateTransactionSchema,
+  }),
+  updateTransactionController,
+);
 
 export default router;